fix: correccionde errores

+++ COMPUESTOS_QUIMICOS_CS_REST_SQL_API/COMPUESTOS_QUIMICOS_CS_REST_SQL_API/Models/Compuesto.cs
@@ -14,7 +14,7 @@ namespace COMPUESTOS_QUIMICOS_CS_REST_SQL_API.Models
         public string? Formula_Quimica { get; set; } = string.Empty;
 
         [JsonPropertyName("masa_molar")]
-        public float Masa_Molar { get; set; } = 0;
+        public decimal Masa_Molar { get; set; } = 0;
 
         [JsonPropertyName("estado_agregacion")]
         public string? Estado_Agregacion { get; set; } = string.Empty;
